Timer role verification
